requirements_ui.txt
streamlit
plotly
tensorflow
pillow
numpy
pandas
matplotlib
seaborn
